requirements.txt
asgiref==3.8.1
Django==4.2.19
django-debug-toolbar==5.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
exceptiongroup==1.2.2
factory_boy==3.3.3
Faker==36.1.1
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytest==8.3.4
pytest-django==4.10.0
python-dotenv==1.0.1
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2025.1
